Partially fixed stacking...again.
Basically, what I'm doing is somewhat parallel to the fixes James has been making on the "master" branch. At the moment, though, the boxes on the tables seem to be working, which is a little farther than what he's got...I've still got a similar problem with the freestanding boxes, though. I'm committing so I can move on to something else in the meantime.