Made Mesh a derivative of AbstractGeometryItem
This required implementing the getUnitHeight() function for meshes, which calculates the distance between the topmost and bottommost points on the mesh (on the y-axis); this will allow meshes to be stacked in the scene graph.

I also moved some of the function implementations from Mesh.h to Mesh.cpp - it's getting a bit crowded in the header file. Some of these are really way too long to be reasonable as inline functions. :-)